requirements.txt
asgiref==3.7.2
Django==5.0.4
sqlparse==0.4.4
psycopg[binary]==3.1.12
djangorestframework==3.15.2